Add timeout/delay for ajax requests and early cancel
This commit adds a `setTimeout` with a small delay around
the `fetch` call. It also moves the outdated request identifier
check into the `fetch` method so that, after the delay/timeout
the request id can be checked against the current request id.
If they are found to be mismatched then a new request for autocomplete
data has since come in and the old one can be safely ignored.

Fixes #9